+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -6,6 +6,9 @@
 
 #include "generated_plugin_registrant.h"
 
+#include <libvips_ffi_windows/libvips_ffi_windows_plugin_c_api.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
+  LibvipsFfiWindowsPluginCApiRegisterWithRegistrar(
+      registry->GetRegistrarForPlugin("LibvipsFfiWindowsPluginCApi"));
 }